% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib.R
\name{contrib_list}
\alias{contrib_list}
\title{Getting the list of the contribution of a contributor.}
\usage{
contrib_list(x, namespace = "0", domain = "fr", page = F)
}
\arguments{
\item{x}{The name of the contributor}

\item{namespace}{The namespace of contributions}

\item{domain}{The domain of the wiki}

\item{page}{If \code{TRUE}, contrib_list return only a list of all different page in witch the contribors made a modification. If \code{FALSE}, return a data-frame with all the contributions, their timestamp and their weight}
}
\value{
Depending the value of page, weither a character vector containing the names of all contributors, or a data-frame containing all the revisions with the name of the contributor, a timestamp and a weight
}
\description{
Getting the list of the contribution of a contributor.
}
\examples{

# All the contribution of an user of the french wiki.
contrib_list('cafeine05') 

# Return a character vector with all the page modified by this contributor.
contrib_list('cafeine05',page=TRUE) 
}

