% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{contribPage}
\alias{contribPage}
\title{Fais la liste des contributeurs d'une page}
\usage{
contribPage(page, domaine = "fr")
}
\arguments{
\item{page}{Le nom de la page à regarder, qui peut appartenir à n'importe quel espace de nom du wiki}
}
\value{
Une liste des contributeurs qui ont contribués à la page
}
\description{
Fais la liste des contributeurs d'une page
}

