% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{contribUser}
\alias{contribUser}
\title{Retourne un vecteurs avec les pages auxquelles le contributeur contribue}
\usage{
contribUser(user)
}
\description{
Retourne un vecteurs avec les pages auxquelles le contributeur contribue
}

