% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus_page_create}
\alias{corpus_page_create}
\title{Creating a page corpus}
\usage{
corpus_page_create(method, x, limit = "max", domain = "fr")
}
\arguments{
\item{method}{The method employed to create the corpus. \describe{
 \item{category}{Every page in the category x is included to the corpus}
 \item{random}{Select x random page}
 \item{xpath}{x is a vector containing in first position the URL of one page,
 and in second position the wpath request to apply to this URL.
 Will return a list of words, wich are the names of pages if the xpath request is writed correctly}
}}

\item{x}{The pointer of the method.}

\item{limit}{If not 'max', then an integer giving the length of the corpus. Useless for random method.}

\item{domain}{The domain where the wiki is located}
}
\value{
A character vector.
}
\description{
Creating a page corpus
}
\examples{
# Creating a page corpus formed by 3 random selected page
corpus_page_create('random',3)
}
\seealso{
\code{\link{corpus_contrib_create}} for contributors corpus, \code{\link{corpus_page_data}} to put data into the corpus you made with this function

Other corpus functions: \code{\link{corpus_contrib_create}},
  \code{\link{corpus_contrib_data}},
  \code{\link{corpus_contrib_select}},
  \code{\link{corpus_page_data}}
}

