% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{createCorpus}
\alias{createCorpus}
\title{Dégage la liste des articles écrits par un groupe de contribteurs}
\usage{
createCorpus(cat, limit = "max", clean = T, threesold = 50, log = T)
}
\arguments{
\item{cat}{La catégorie où se trouvent les pages utilisateurs concernées}
}
\value{
Un vecteur d'article
}
\description{
Dégage la liste des articles écrits par un groupe de contribteurs
}

