% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{dataTable}
\alias{dataTable}
\title{Construit un tableau de données à partir d'un corpus}
\usage{
dataTable(corpus, selection = c("nbLinks", "nbContrib", "nbRevisions",
  "percentAnon", "percent10"))
}
\description{
Construit un tableau de données à partir d'un corpus
}

