% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{divideList}
\alias{divideList}
\title{Divise une liste en vecteur d'une certaine taille}
\usage{
divideList(x, number)
}
\arguments{
\item{x}{La liste a divisé}

\item{number}{La longueur des divisions}
}
\value{
Une liste de vecteur de longueur number décomposant la liste x
}
\description{
Divise une liste en vecteur d'une certaine taille
}

