% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph.R
\name{graphArticle}
\alias{graphArticle}
\title{Make the graph of the links of one page}
\usage{
graphArticle(titre, domaine = "fr", namespace = "0")
}
\arguments{
\item{title}{the title of a page}

\item{domain}{the wiki of a page}
}
\value{
A graph object
}
\description{
Make the graph of the links of one page
}

