% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graphArticleGroup}
\alias{graphArticleGroup}
\title{Réalise le graph d'un groupe d'articles}
\usage{
graphArticleGroup(group, domaine = "fr", namespace = "0")
}
\arguments{
\item{group}{Le groupe d'article à grapher, qui sera aussi la liste des noeuds du graphe}
}
\value{
Un objet igraph
}
\description{
Réalise le graph d'un groupe d'articles
}

