% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{keepSimilar}
\alias{keepSimilar}
\title{Compare deux vecteurs}
\usage{
keepSimilar(vector1, vector2)
}
\arguments{
\item{vector1}{Le premier vecteur}

\item{vector2}{Le second vecteur}
}
\value{
Un vecteur ne contenant que les valeurs communes
}
\description{
Compare deux vecteurs
}

