% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{links}
\alias{links}
\title{Download a list of all the links in a Wikipedia page}
\usage{
links(titre, domaine = "fr", namespace = "0")
}
\arguments{
\item{domaine}{The domain of the wiki. For example, for the english wikipedia, the domain is "en" (because the adress of the whole website is http://en.wikipedia.org/)}

\item{title}{The title of the page}
}
\value{
A list of all links of a page
}
\description{
Download a list of all the links in a Wikipedia page
}

