% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infos.R
\name{nbLinks}
\alias{nbLinks}
\title{Nombre de liens d'une page}
\usage{
nbLinks(page, domaine = "fr", namespace = "0")
}
\arguments{
\item{page}{Le titre de la page}

\item{namespace}{Les namespace vers lesquels doivent pointer la page. Par défaut 0, celui des articles.}
}
\value{
Un vecteur numérique de dimension 1
}
\description{
Nombre de liens d'une page
}

