% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_anon}
\alias{page_anon}
\title{Giving the proportion of anonmous contribution in a revisions list}
\usage{
page_anon(revisions)
}
\arguments{
\item{revisions}{Revision list built with page_revisions}
}
\value{
An integer.
}
\description{
Giving the proportion of anonmous contribution in a revisions list
}
\examples{
# Downloading the list of the revisions of the article 'Action' in the french wiki
revisions <- page_revisions('Action') 

# Returning the percentage of anonymous contribution in this article
page_anon(revisions) 
}
\seealso{
Other page functions: \code{\link{page_infos}},
  \code{\link{page_islink}}, \code{\link{page_links}},
  \code{\link{page_ranking}}, \code{\link{page_revisions}}
}

