% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_cat}
\alias{page_cat}
\title{List all categories where the page is located}
\usage{
page_cat(title, domain = "fr")
}
\arguments{
\item{title}{The title of the page}

\item{domain}{The domain where the wiki is located}
}
\value{
A character vector giving
}
\description{
List all categories where the page is located
}
\examples{
# Downloading the list of categories in the action page
page_cat('Action') 
}

