% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_infos}
\alias{page_infos}
\title{Getting some quantitative infos about a wiki page}
\usage{
page_infos(title, domain = "fr")
}
\arguments{
\item{title}{The title of the page}

\item{domain}{The domain where the wiki is located}
}
\value{
A named list
}
\description{
Getting some quantitative infos about a wiki page
}
\examples{
# Getting a list of information for the 'Nationalism' page in the french wiki
page_infos('Nationalisme') 
}
\seealso{
Other page functions: \code{\link{page_anon}},
  \code{\link{page_islink}}, \code{\link{page_links}},
  \code{\link{page_ranking}}, \code{\link{page_revisions}}
}

