% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_islink}
\alias{page_islink}
\title{Test if page(s) is(are) linked}
\usage{
page_islink(from, to, namespace = "0", domain = "fr")
}
\arguments{
\item{from}{The title of the page from the link is supposed to go from}

\item{to}{Either a string or a character vector representing the page(s) where the link(s) is(are) supposed to go to.}

\item{namespace}{The namespace of the page to look.}

\item{domain}{The domain where the wiki is located}
}
\value{
A character vector.
}
\description{
Test if page(s) is(are) linked
}
\examples{
# Testing if the article 'Action' contains a link to 'Sociologie' in the french wiki.
page_islink('Action','Sociologie') 
}
\seealso{
Other page functions: \code{\link{page_anon}},
  \code{\link{page_infos}}, \code{\link{page_links}},
  \code{\link{page_ranking}}, \code{\link{page_revisions}}
}

