% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_links}
\alias{page_links}
\title{List of inter-articles links on a wiki page}
\usage{
page_links(x, domain = "fr", namespace = "0")
}
\arguments{
\item{x}{The title of the page}

\item{domain}{The domain of the wiki where the page is located}

\item{namespace}{The namespace where the page need to be to be ept in the graph}
}
\value{
A character vector
}
\description{
List of inter-articles links on a wiki page
}
\examples{
page_links('Action') # listing all the links contained in the 'action' article.
}
\seealso{
Other page functions: \code{\link{page_anon}},
  \code{\link{page_infos}}, \code{\link{page_islink}},
  \code{\link{page_ranking}}, \code{\link{page_revisions}}
}

