% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_ranking}
\alias{page_ranking}
\title{Extracting the first contributors freom a list of revisions}
\usage{
page_ranking(revisions, threesold, keepAnon = TRUE)
}
\arguments{
\item{revisions}{A list of revisions built with page_revisions}

\item{threesold}{An integer, he function will return the \code{threesold} percent first contributors}

\item{keepAnon}{A boolean indicate weither the anonymous should be kept into the list or not}
}
\value{
A character vector
}
\description{
Extracting the first contributors freom a list of revisions
}
\examples{
# Downloading the list of revisions
revisions <- page_revisions('Action')

# Extracting the list of the 10 percent first contributor of the list
page_ranking(revisions,10,TRUE) 
}
\seealso{
Other page functions: \code{\link{page_anon}},
  \code{\link{page_infos}}, \code{\link{page_islink}},
  \code{\link{page_links}}, \code{\link{page_revisions}}
}

