% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_revisions}
\alias{page_revisions}
\title{Downloading the list of contributions for one page}
\usage{
page_revisions(page, domain = "fr", contrib = F)
}
\arguments{
\item{page}{The title of the page}

\item{domain}{The domain where the wiki is located}

\item{contrib}{A boolean indicating whether to return or not only contributors' names}
}
\value{
A data-frame containing the username of the user (or the IP if anonymous contribution), the timestamp, the size of the revision, a boolean indicating weither the contribution is anonymous or not, and the difference beetween the contribution and the previous
}
\description{
Downloading the list of contributions for one page
}
\examples{
# Downloading the list of contribution for the 'action' page in the french wiki
page_revisions('Action') 
}
\seealso{
Other page functions: \code{\link{page_anon}},
  \code{\link{page_infos}}, \code{\link{page_islink}},
  \code{\link{page_links}}, \code{\link{page_ranking}}
}

