% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{pageCat}
\alias{pageCat}
\title{Lister les catégories auxquelles appartient une page}
\usage{
pageCat(titre, domaine = "fr")
}
\arguments{
\item{titre}{Le titre de la page}

\item{domaine}{Le domaine désignant le wiki.}
}
\value{
Un vecteur de caractères listant les catégories.
}
\description{
Lister les catégories auxquelles appartient une page
}

