% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infos.R
\name{percentAnon}
\alias{percentAnon}
\title{Retourne le pourcentage de contribution anonymes d'un article}
\usage{
percentAnon(list)
}
\arguments{
\item{list}{Une liste de contribution construite avec revisionsPage}
}
\value{
Un pourcentage de contribution
}
\description{
Retourne le pourcentage de contribution anonymes d'un article
}

