% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{period}
\alias{period}
\title{Recode a date vector into periods}
\usage{
period(date, period = c("2005-05-01", "2009-12-01"))
}
\arguments{
\item{date}{A timestamp vector}

\item{period}{The breaks witch \code{date} sould be cut with}
}
\value{
a numeric vector with the periods replacing the date
}
\description{
Recode a date vector into periods
}

