% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infos.R
\name{rankingContrib}
\alias{rankingContrib}
\title{Range les contributeurs en fonction du poids de leur contribution}
\usage{
rankingContrib(list, threesold, keepAnon = TRUE)
}
\arguments{
\item{list}{Une liste de révision produite par revisionsPage}

\item{threesold}{Un seuil exprimé en pourcentage de contribution dans le poids de l'article pour garder le contributeur dans la liste}

\item{keepAnon}{Une valeur booléenne indiquant si oui ou non on doit conserver les contributeurs anonymes dans le classement}
}
\value{
Une liste comprennant 1) un vecteur classant les contributeurs entre eux. 2) le poids des contributeurs présent dans le vecteur, en % de contribution à la page
}
\description{
Range les contributeurs en fonction du poids de leur contribution
}

