% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{revisionsPage}
\alias{revisionsPage}
\title{Une liste de toutes les révisions d'une page}
\usage{
revisionsPage(page, domaine = "fr")
}
\arguments{
\item{page}{Le titre de la page}
}
\value{
Un data-frame avec 5 colonnes : 1) l'utilisateur qui a fait la révision, 2) la date de la révision (sous forme de timestamp), 3) la taille en octet de la révision, 4) une valeur booléenne indicant si, oui ou non, la revision est faite par un utilisateur anonyme, 5) la taille de la révision
}
\description{
Une liste de toutes les révisions d'une page
}

