% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{selectRandom}
\alias{selectRandom}
\title{Selectionner un corpus d'articles de manière aléatoire.}
\usage{
selectRandom(nb, domaine = "fr", namespace = "0")
}
\arguments{
\item{nb}{Le nombre d'article à obtenir}

\item{domaine}{le domaine du wiki de Wikipédia où l'on veut sélectionner les données}

\item{namespace}{le namespace des articles}
}
\description{
Selectionner un corpus d'articles de manière aléatoire.
}

