% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infos.R
\name{startPage}
\alias{startPage}
\title{Obtenir la première révision d'une page}
\usage{
startPage(page, domaine = "fr")
}
\arguments{
\item{page}{le titre de la page}

\item{domaine}{le domaine où est localisé le wiki}
}
\value{
un data-frame à une ligne contenant la date, l'utilisateur et la taille de la révision
}
\description{
Obtenir la première révision d'une page
}

