% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trajectory.R
\name{user}
\alias{user}
\title{Download a list of all the contributions of an user}
\usage{
user(user, namespace = "0", domaine = "fr", traminer = FALSE)
}
\description{
Download a list of all the contributions of an user
}

