% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{xPath}
\alias{xPath}
\title{Executer une requête xPath sur une page}
\usage{
xPath(url, xpath)
}
\arguments{
\item{url}{l'URL de la page. Attention : différent du titre de la page, il faut l'URL complète !}

\item{xpath}{la requête à exécuter.}
}
\value{
un vecteur texte
}
\description{
Executer une requête xPath sur une page
}

