% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graphArticle}
\alias{graphArticle}
\title{Réalise le graphe d'une page}
\usage{
graphArticle(titre, domaine = "fr", namespace = "0")
}
\arguments{
\item{title}{Le titre de la page}

\item{domain}{Le domaine du wiki}
}
\value{
Un objet igraph
}
\description{
Réalise le graphe d'une page
}

