% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.R
\name{keepSimilar}
\alias{keepSimilar}
\title{Retourne un vecteur qui ne contient que les valeurs communes aux deux arguments.}
\usage{
keepSimilar(vector1, vector2)
}
\description{
Retourne un vecteur qui ne contient que les valeurs communes aux deux arguments.
}

