% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{links}
\alias{links}
\title{Trouver l'ensemble des liens d'une page}
\usage{
links(titre, domaine = "fr", namespace = "0")
}
\arguments{
\item{title}{Le titre de la page}
}
\value{
Une liste de tous les liens de la page
}
\description{
Trouver l'ensemble des liens d'une page
}

