% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{user}
\alias{user}
\title{Liste des contributions d'un utilisateur}
\usage{
user(user, namespace = "0", domaine = "fr", date = TRUE, weight = TRUE)
}
\arguments{
\item{user}{Le pseudonyme de l'utilisateur}

\item{namespace}{L'espace de nom dont on veut lister les pages, par défaut 0 correspondant à tous les articles}

\item{domaine}{Le domaine du wiki}

\item{date}{Une valeur booléenne qui indique si oui ou non on veut les dates des contributions}

\item{weight}{Une valeur booléenne qui indique si oui ou non on veut le poids de la contribution (ralonge beaucoup le temps de calcul)}
}
\value{
Un data-frame qui contient les données spécifiées par les paramètres
}
\description{
Liste des contributions d'un utilisateur
}

