% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{xpath}
\alias{xpath}
\title{Query a page by xpath means}
\usage{
xpath(url, xpath)
}
\arguments{
\item{url}{URL of the page. Please note that this is different of the title of the wiki page.}

\item{xpath}{XPATH query to execute}
}
\value{
A character vector
}
\description{
Query a page by xpath means
}

